LayoutTests/imported/w3c/web-platform-tests/css/motion/offset-supports-calc-expected.txt


PASS offset-position supports calc
FAIL offset-path supports calc assert_equals: expected (string) "ray(270deg closest-side)" but got (undefined) undefined
PASS offset-distance supports calc
FAIL offset-rotate supports calc assert_equals: expected (string) "auto 270deg" but got (undefined) undefined
PASS offset-anchor supports calc
